Migrate server subproject tests to munit suites
